add stiles for categories
