+++ 01_React/money-pal/src/components/Categories/Categories.css
@@ -1,0 +1,107 @@
+
+.categories__container {
+    display: flex;
+    width: 100%;
+    justify-content: center;
+}
+
+.categories__container > div {
+    width: 15%;
+}
+
+.category {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+.category p {
+    margin: 0 auto;
+    font-size: 12px;
+}
+
+.category__input {
+    width: 40px;
+    height: 40px;
+    position: relative;
+    border-radius: 50%;
+}
+
+.category__input span {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    display: block;
+    background: #d5d6cf;
+    border-radius: 50%;
+}
+.category__input input {
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    top: 0;
+    left: 0;
+    display: block;
+    opacity: 0;
+}
+
+.category__input input:checked ~ span {
+    background-position: 50% 50%;
+    background-size: contain;
+}
+/*transport*/
+.category--transport input ~ span {
+    background: url("./img/transport.svg") #9cc3a1 50% 50% no-repeat;
+    background-size: 65%;
+}
+.category--transport input:checked ~ span {
+    background-size: 70%;
+    background-color: #56c387;
+}
+/*food*/
+.category--food input ~ span {
+    background: url("./img/food.svg") #c38c90 50% 50% no-repeat;
+    background-size: 62%;
+}
+.category--food input:checked ~ span {
+    background-size: 68%;
+    background-color: #c32150;
+}
+/*clothes*/
+.category--clothes input ~ span {
+    background: url("./img/clothes.svg") #8298c3 50% 50% no-repeat;
+    background-size: 65%;
+}
+.category--clothes input:checked ~ span {
+    background-size: 70%;
+    background-color: #416ec3;
+}
+/*entertainment*/
+.category--entertainment input ~ span {
+    background: url("./img/entertainment.svg") #dfaa86 50% 50% no-repeat;
+    background-size: 65%;
+}
+.category--entertainment input:checked ~ span {
+    background-size: 70%;
+    background-color: #df8748;
+}
+/*bills*/
+.category--bills input ~ span {
+    background: url("./img/bills.svg") #7ec0c3 50% 50% no-repeat;
+    background-size: 65%;
+}
+.category--bills input:checked ~ span {
+    background-size: 70%;
+    background-color: #3dc3c2;
+}
+/*other*/
+.category--other input ~ span {
+    background: url("./img/other.svg") #aa95c3 50% 50% no-repeat;
+    background-size: 65%;
+}
+.category--other input:checked ~ span {
+    background-size: 70%;
+    background-color: #901ac3;
+}
